Move domain delegation to new files
