+++ src/pkg/cli/client/byoc/aws/byoc_integration_test.go
@@ -8,8 +8,6 @@ import (
 	"testing"
 
 	defangv1 "github.com/DefangLabs/defang/src/protos/io/defang/v1"
-	"github.com/aws/aws-sdk-go-v2/config"
-	"github.com/aws/aws-sdk-go-v2/service/route53"
 	"github.com/bufbuild/connect-go"
 )
 
@@ -151,17 +149,4 @@ func TestListSecrets(t *testing.T) {
 			t.Fatalf("expected empty list, got %v", secrets.Names)
 		}
 	})
-}
-
-func TestPrepareDomainDelegation(t *testing.T) {
-	ctx := context.Background()
-	cfg, err := config.LoadDefaultConfig(ctx)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	r53Client := route53.NewFromConfig(cfg)
-
-	testPrepareDomainDelegationNew(t, r53Client)
-	testPrepareDomainDelegationLegacy(t, r53Client)
 }
